+++ src/index.js
@@ -1,102 +1,63 @@
-import 'babel-polyfill';
+import puppeteer from 'puppeteer';
 import fs from 'fs';
-import Chromy from 'chromy';
-import getPort from 'get-port';
+import chainProxy from './helpers/proxyChain';
 import { sanitiseGlobalConfiguration, sanitiseTestConfiguration } from './sanitiser';
-import run from './chromyRunner';
-import logger from './logger';
-import { configTypes } from './defaultConfig';
-import actions from './actions';
+import Page from './page';
+import logger from './utils/logger';
 
-const CHROME_WIDTH = 800;
-const CHROME_HEIGHT = 600;
-
-const createDir = (path) => {
-  const screentshotsPath = `${path}`;
-  if (!fs.existsSync(screentshotsPath)) {
-    fs.mkdirSync(screentshotsPath);
-  }
-};
-
-const getFreePort = async () => {
-  try {
-    return await getPort();
-  } catch (error) {
-    logger.error('Failed to get a free port', error);
-    return null;
-  }
+const createDirs = (paths) => {
+  paths.forEach((path) => {
+    if (!fs.existsSync(path)) {
+      fs.mkdirSync(path);
+    }
+  });
 };
 
 export default class Differencify {
   constructor(conf) {
     this.configuration = sanitiseGlobalConfiguration(conf);
-    this.chromeInstances = {};
+    this.browser = null;
+    this.testId = 0;
     if (this.configuration.debug === true) {
       logger.enable();
     }
-    createDir(this.configuration.screenshots);
-    createDir(this.configuration.testReportPath);
-  }
-
-  async _createChromeInstance(testConfig) {
-    const width = testConfig.resolution.width || CHROME_WIDTH;
-    const height = testConfig.resolution.height || CHROME_HEIGHT;
-    const flags = [`--window-size=${width},${height}`];
-    const port = await getFreePort();
-    if (!port) {
-      return null;
+    if (this.configuration.isUpdate) {
+      logger.warn('Your tests are running on update mode. Your screenshots will be updated');
     }
-    const chromy = new Chromy({
-      port,
-      chromeFlags: flags,
-      waitTimeout: this.configuration.timeout,
-      visible: this.configuration.visible,
-    });
-    return chromy;
+    createDirs([this.configuration.screenshots, this.configuration.testReports]);
   }
 
-  _updateChromeInstances(chromy) {
-    this.chromeInstances[chromy.options.port] = chromy;
-  }
-
-  async _closeChrome(chromy, testName) {
-    try {
-      logger.prefix(testName).log('closing browser');
-      await chromy.close();
-      delete this.chromeInstances[chromy.options.port];
-    } catch (error) {
-      logger.prefix(testName).log(error);
+  async launchBrowser() {
+    if (!this.browser) {
+      logger.log('Launching browser...');
+      try {
+        this.browser = await puppeteer.launch(this.configuration.puppeteer);
+      } catch (error) {
+        logger.error(error);
+      }
+    } else {
+      logger.log('Using existing browser instance');
     }
   }
 
-  async _run(config, type) {
-    const testConfig = sanitiseTestConfiguration(config);
-    const chromy = await this._createChromeInstance(testConfig);
-    if (!chromy) {
-      return false;
+  init(testConfig) {
+    this.testId += 1;
+    const config = sanitiseTestConfiguration(testConfig, this.testId);
+    logger.prefix(config.testName).log('Opening new tab...');
+    const page = new Page(this.browser, config, this.configuration);
+    if (config.chain) {
+      return chainProxy(page, this.configuration.isUpdate);
     }
-    this._updateChromeInstances(chromy);
-    testConfig.type = type;
-    const result = await run(chromy, this.configuration, testConfig);
-    await this._closeChrome(chromy, testConfig.name);
-    return result;
-  }
-
-  async update(config) {
-    return await this._run(config, configTypes.update);
-  }
-
-  async test(config) {
-    config.steps.push({ name: actions.test, value: this.configuration.testReportPath });
-    return await this._run(config, configTypes.test);
+    return page;
   }
 
   async cleanup() {
-    await Promise.all(
-      Object.values(this.chromeInstances).map(chromeInstance => chromeInstance.close()),
-    );
-    this.chromeInstances = {};
-    logger.log('All browsers been closed');
+    logger.log('Closing browser...');
+    try {
+      await this.browser.close();
+    } catch (error) {
+      logger.error(error);
+    }
   }
 }