+++ src/sanitiser.test.js
@@ -4,60 +4,21 @@ describe('sanitiser', () => {
   describe('sanitise global configuration', () => {
     it('sanitise if no config provided', () => {
       const configuration = sanitiseGlobalConfiguration({});
-      expect(configuration).toEqual({
-        screenshots: './screenshots',
-        testReportPath: './differencify_report',
-        saveDifferencifiedImage: true,
-        debug: false,
-        visible: false,
-        timeout: 30000,
-        mismatchThreshold: 0.01,
-      });
+      expect(configuration).toMatchSnapshot();
     });
     it('sanitise if screenshots config provided', () => {
-      const configuration = sanitiseGlobalConfiguration({ screenshots: './screenshots' });
-      expect(configuration).toEqual({
-        screenshots: './screenshots',
-        testReportPath: './differencify_report',
-        saveDifferencifiedImage: true,
-        debug: false,
-        visible: false,
-        timeout: 30000,
-        mismatchThreshold: 0.01,
-      });
+      const configuration = sanitiseGlobalConfiguration({ screenshots: './somewhere' });
+      expect(configuration).toMatchSnapshot();
     });
   });
   describe('sanitise test configuration', () => {
     it('sanitise if no config provided', () => {
-      const configuration = sanitiseTestConfiguration({});
-      expect(configuration).toEqual({
-        name: 'default',
-        resolution: {
-          width: 800,
-          height: 600,
-        },
-        steps: [
-          { name: 'goto', value: 'www.example.com' },
-          { name: 'capture', value: 'document' },
-        ],
-      });
+      const configuration = sanitiseTestConfiguration({}, 1);
+      expect(configuration).toMatchSnapshot();
     });
-    it('sanitise if resolution config provided', () => {
-      const configuration = sanitiseTestConfiguration({ resolution: {
-        width: 1366,
-        height: 768,
-      } });
-      expect(configuration).toEqual({
-        name: 'default',
-        resolution: {
-          width: 1366,
-          height: 768,
-        },
-        steps: [
-          { name: 'goto', value: 'www.example.com' },
-          { name: 'capture', value: 'document' },
-        ],
-      });
+    it('sanitise if some config provided', () => {
+      const configuration = sanitiseTestConfiguration({ newWindow: true }, 1);
+      expect(configuration).toMatchSnapshot();
     });
   });
 });